Removed import from book display and moved functions into same file.
