+++ contents/euclidean_algorithm/code/kotlin/Euclidean.kt
@@ -12,6 +12,20 @@ fun euclidSub(a: Int, b: Int): Int {
     return a
 }
 
+fun euclidMod(a: Int, b: Int): Int {
+    var a = a.absoluteValue
+    var b = b.absoluteValue
+
+    while (b != 0) {
+        val tmp = b
+        b = a % b
+        a = tmp
+    }
+
+    return a
+}
+
 fun main() {
     println(euclidSub(128 * 12, 128 * 77))
+    println(euclidMod(64 * 67, 64 * 81))
 }
