--- contents/euclidean_algorithm/code/kotlin/EuclideanMod.kt
@@ -1,18 +1,0 @@
-import kotlin.math.absoluteValue
-
-fun euclidMod(a: Int, b: Int): Int {
-    var a = a.absoluteValue
-    var b = b.absoluteValue
-
-    while (b != 0) {
-        val tmp = b
-        b = a % b
-        a = tmp
-    }
-
-    return a
-}
-
-fun main(args: Array<String>) {
-    println(euclidMod(64 * 67, 64 * 81))
-}
