+++ contents/euclidean_algorithm/euclidean_algorithm.md
@@ -16,7 +16,7 @@ The algorithm is a simple way to find the *greatest common divisor* (GCD) of two
 {% sample lang="java" %}
 [import:3-16, lang="java"](code/java/EuclideanAlgo.java)
 {% sample lang="kotlin" %}
-[import:1-13, lang="kotlin"](code/kotlin/EuclideanSub.kt)
+[import:3-13, lang="kotlin"](code/kotlin/Euclidean.kt)
 {% sample lang="js" %}
 [import:15-29, lang="javascript"](code/javascript/euclidean_example.js)
 {% sample lang="lisp" %}
@@ -89,7 +89,7 @@ Modern implementations, though, often use the modulus operator (%) like so
 {% sample lang="java" %}
 [import:18-26, lang="java"](code/java/EuclideanAlgo.java)
 {% sample lang="kotlin" %}
-[import:1-14, lang="kotlin"](code/kotlin/EuclideanMod.kt)
+[import:15-26, lang="kotlin"](code/kotlin/Euclidean.kt)
 {% sample lang="js" %}
 [import:1-13, lang="javascript"](code/javascript/euclidean_example.js)
 {% sample lang="lisp" %}
